refactor: extract cBioPortal ingest and slice YAMLs to showcase/cbioportal_to_omop/
